move HiveProvider to shims
